+++ examples/src/tests/holo.rs
@@ -27,7 +27,7 @@ pub async fn holo(autd: &mut Controller<impl Link>) -> anyhow::Result<bool> {
 
     let backend = NalgebraBackend::new()?;
 
-    let target_amp = 2.5e3 * autd.geometry.num_devices() as float * Pascal;
+    let target_amp = 2.5e3 * autd.geometry.num_devices() as f64 * Pascal;
     match s.trim().parse::<usize>() {
         Ok(0) => {
             let g = SDP::new(backend)
